Extract duplicated code into trait
