+++ src/PaymentMethods/PaymentFieldsStrategies/RivertyFieldsStrategy.php
@@ -6,6 +6,7 @@ namespace Mollie\WooCommerce\PaymentMethods\PaymentFieldsStrategies;
 
 class RivertyFieldsStrategy implements PaymentFieldsStrategyI
 {
+    use PaymentFieldsStrategiesTrait;
     const FIELD_BIRTHDATE = "billing_birthdate";
     const FIELD_PHONE = "billing_phone_riverty";
 
@@ -39,28 +40,6 @@ class RivertyFieldsStrategy implements PaymentFieldsStrategyI
         if ($showBirthdateField) {
             $this->dateOfBirth($birthValue);
         }
-    }
-
-    protected function getOrderIdOnPayForOrderPage()
-    {
-        global $wp;
-        $orderId = absint($wp->query_vars['order-pay']);
-        return wc_get_order($orderId);
-    }
-
-    protected function dateOfBirth($birthValue)
-    {
-        $birthValue = $birthValue ?: '';
-        ?>
-        <p class="form-row form-row-wide" id="billing_birthdate_field">
-            <label for="<?php echo esc_attr(self::FIELD_BIRTHDATE); ?>" class=""><?php echo esc_html__('Birthdate', 'mollie-payments-for-woocommerce'); ?>
-            </label>
-            <span class="woocommerce-input-wrapper">
-                <input type="date" class="input-text " name="<?php echo esc_attr(self::FIELD_BIRTHDATE); ?>"
-                       id="<?php echo esc_attr(self::FIELD_BIRTHDATE); ?>" value="<?php echo esc_attr($birthValue); ?>"
-                       autocomplete="birthdate"></span>
-        </p>
-        <?php
     }
 
     protected function phoneNumber($phoneValue)
